+++ packages/datatrak-web-server/src/app/PostgresLoadTester.ts
@@ -1,21 +1,12 @@
 // @ts-nocheck
-import { PGlite } from '@electric-sql/pglite';
-import { CREATE_TABLES_SCRIPT, TABLE_INSERT_QUERIES, TABLES } from './loadTestingQueries';
-import generateId from 'uuid/v1';
+// All errors in this file will be ignored
+import { TupaiaDatabase, generateId } from '@tupaia/database';
+import { INSERT_SURVEY, SURVEY_ID, TABLE_INSERT_QUERIES, TABLES } from './loadTestingQueries';
 
-class PGLiteLoadTester {
+class PostgresLoadTester {
   constructor() {
-    this.db = new PGlite('idb://my-pgdata');
+    this.db = new TupaiaDatabase();
     this.results = { reads: [], writes: [], concurrentOps: [], errors: [] };
-  }
-
-  async initTestTable() {
-    try {
-      await this.db.exec(CREATE_TABLES_SCRIPT);
-    } catch (error) {
-      console.error(error);
-    } finally {
-    }
   }
 
   async testWrite(table, totalOps, parameters = [], parameterGenerator = () => []) {
@@ -23,7 +14,7 @@ class PGLiteLoadTester {
     try {
       for (let i = 0; i < totalOps; i++) {
         const queryParameters = parameters.length ? parameters : parameterGenerator(i);
-        await this.db.query(TABLE_INSERT_QUERIES[table], [generateId(), ...queryParameters]);
+        await this.db.executeSql(TABLE_INSERT_QUERIES[table], [generateId(), ...queryParameters]);
       }
     } catch (error) {
       this.results.errors.push({ table, operation: 'write', error: error.message });
@@ -42,10 +33,10 @@ class PGLiteLoadTester {
 
     try {
       for (let i = 0; i < totalOps; i++) {
-        await this.db.query(
+        await this.db.executeSql(
           `
           SELECT * FROM ${table}
-          LIMIT $1
+          LIMIT ?
         `,
           [limit],
         );
@@ -64,20 +55,17 @@ class PGLiteLoadTester {
   }
 
   async runFullTest() {
-    console.log('Initializing test table...');
-    await this.initTestTable();
-
     const surveyId = generateId();
 
-    await this.db.query(TABLE_INSERT_QUERIES[TABLES.SURVEY], [surveyId, surveyId, surveyId]);
+    await this.db.executeSql(INSERT_SURVEY, [surveyId, surveyId, surveyId]);
 
     console.log('Testing write data...');
     await this.testWrite(TABLES.SURVEY_RESPONSE, 1000, [surveyId]);
 
-    const surveyResponseResults = await this.db.query(
+    const surveyResponses = await this.db.executeSql(
       `
         SELECT * FROM survey_response
-        WHERE survey_id = $1
+        WHERE survey_id = ?
         ORDER BY data_time
         LIMIT 1000;
       `,
@@ -86,17 +74,17 @@ class PGLiteLoadTester {
 
     await this.testWrite(TABLES.DATA_ELEMENT, 1000);
 
-    const dataElementResults = await this.db.query(
+    const [dataElement] = await this.db.executeSql(
       `
         SELECT * FROM data_element 
         LIMIT 1;
       `,
     );
-    const dataElementId = dataElementResults.rows[0].id;
+    const dataElementId = dataElement.id;
 
     await this.testWrite(TABLES.QUESTION, 1000, [dataElementId]);
 
-    const questionResults = await this.db.query(
+    const questions = await this.db.executeSql(
       `
         SELECT * FROM question 
         LIMIT 1000;
@@ -112,7 +100,7 @@ class PGLiteLoadTester {
       }
 
       questionsIndex++;
-      return [surveyResponseResults.rows[surveyResponsesIndex].id, questionResults.rows[questionsIndex].id];
+      return [surveyResponses[surveyResponsesIndex].id, questions[questionsIndex].id];
     };
 
     await this.testWrite(TABLES.ANSWER, 1000, [], answerParametersGenerator);
@@ -123,15 +111,13 @@ class PGLiteLoadTester {
     await this.testRead(TABLES.QUESTION, 1000);
     await this.testRead(TABLES.ANSWER, 1000);
 
-    console.log('Test concurrent ops...');
-
     return this.results;
   }
 }
 
 // Example usage:
 export const runLoadTest = async () => {
-  const tester = new PGLiteLoadTester();
+  const tester = new PostgresLoadTester();
 
   try {
     const results = await tester.runFullTest();
