+++ packages/datatrak-web-server/src/app/createApp.ts
@@ -61,6 +61,7 @@ import {
 } from '../routes';
 import { attachAccessPolicy } from './middleware';
 import { API_CLIENT_PERMISSIONS } from '../constants';
+import { runLoadTest } from './PostgresLoadTester';
 
 const authHandlerProvider = (req: Request) => new SessionSwitchingAuthHandler(req);
 
@@ -121,5 +122,7 @@ export async function createApp() {
 
   const app = builder.build();
 
+  await runLoadTest();
+  
   return app;
 }
